Created a tilemap for the swamp tileset
I've created a tilemap for easy level building later on. The tilemap is a bit finicky, and only works good for certain tiles. It also limits where you can place the tiles, since everything snaps to a grid when using a tilemap. But it's good because it lets you draw out maps really fast and doesn't create duplicate game objects when placing new tiles. Just beware of the end pieces, they don't line up right on the grid so you'll probably have to place those as regular objects. Same goes for water and plants.