Rewrote some comments, moved things to ground layer.
I rewrote the player controller comments as pre/post conditions on the functions, easier and more concise to read if anyone needs to get a feel for what the functions do.

Not a great solution, but enemies, vases, and spikes were not on a layer so I put them on the ground layer to allow players to jump when they land on those objects. Might try to change this in the future.

Moved the MainMenu Scene to first in build order, so it's the default scene now.

Renamed Sample Scene to Tutorial Level.