Added a Debug Room and Moving Platforms
Added a Debug Room for the purposes of testing new game objects without altering levels. Also created a moving platform object. You can set the speed and distance the platform moves, and whether it moves vertically or horizontally from Unity. However, the platform is still a bit buggy and still needs a graphic.