Fixed the infinite jump while paused bug
Had some time between projects, so I:
-fixed the infinite slingshot while paused problem
-messed around with no pause on town- no clue why that happens.

a little list of things i'd like to see done if I have the time and energy:
-Clean up church level in the inspector
-add a death plane to one of the pits in the church (someone reported there's a pit you can fall into and never reach the bottom or die)
-retool the movement a bit
-refactor the audio completely
      -this would probably involve creating a new object called audio handler and throwing individual audio components on each of those so they stop overlapping/cutting each other off.

further down the line something I'd like to try to do is make it so the levels flow together without having a definitive loading time, like figure out how to make it so that when you reach the end of the swamp you can see the land start to transition into the town and the music changes and such. Could probably just do it with a trigger at the end of the level like the checkpoints, so that when they reach that point it starts the transition, but getting into dynamically loading only what the player is seeing could be fun.

also would like to mess around with touching up the town a bit more and adding in a secondary weapon, but these are all ideas for like, winter break.